FIX: renaming of files with no numbers
Titles *should* have the episode number in them. But if the do not and
there is no number, they get moved to a file with no filename and only
an extension. This commit fixes that.

Also some output has been changed to make it more readable and `mv` is
only executed if `$f` and `$fNewName` are not equal.