Skip stripping letters from the previews dir
Since these will normally be named after the title rather than episode
numbers. This directory will not strip letters from filenames. However,
an MD5 sum will be added to the end of the filename. This way if the
file changes or one of the same title is added later, they won't overwrite
eachother.